Modification to Caching Strategy
The application will no longer maintain a seperate set of repositories
that interface with our caching mechanism, in this case Redis. Instead
we have a single service that exports a Service defined interface
that defines all of the supported functionality. All other services
that need access to the cache can import this service and call the functions
necessary to retrieve the data they need.