New Services and Repositories
Created Alliance, Corporation, and Character Services and their
subsequent respositories. There is no much there, but the pattern
has been established. Modified signature of Create and Update to acommodate
id string which should be the MongoDB ObjectID is string format. The repo
with take this in and convert it to an actual ObjectID